Print out the target speed in waypoint loader to debug the issue reported by Udacity review
Looking at the video record, I think the reviewer's conclude that
"The car remained stationary for a long time even though it was supposed to move."
is certainly wrong. The car did move, but the speed was capped at 7-8 mph.
I suspect the rosparam for target velocity was modified (our codebase set it to 40 mph).
Add this print out on console and resubmit.